# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SMS Gateway",
    "version": "1.0.0",
    "description": "HTTP API that sends and receives SMS through one or more USB GSM modems (serial AT, ZTE HiLink HTTP, or Android/adb drivers). Sends go through a Redis-backed queue with automatic retries; without a pinned `modem`, jobs are spread round-robin across connected modems.\n\n**Message limits** — a message longer than one SMS is sent as a concatenated SMS of up to 3 parts:\n\n| Driver | Charset | 1 part | max (3 parts) |\n|---|---|---|---|\n| serial | GSM-7 (plain Latin) | 160 | 459 (153/part) |\n| serial | UCS2 (e.g. Cyrillic) | 70 | 201 (67/part) |\n| zte-http | always UCS2 | 70 | 201 (67/part) |\n| adb | GSM-7 / UCS2 (auto) | 160 / 70 | 459 / 201 |\n\nWithout a pinned `modem`, the message must fit every configured modem's limits.\n\n**Carrier content filtering** — some operators (observed with TMCell) silently reject messages containing `http`; avoid URLs in message bodies."
  },
  "servers": [{ "url": "/" }],
  "security": [{ "ApiKeyAuth": [] }, { "BearerAuth": [] }],
...
  "paths": {
...
    "/sms/sent": {
      "get": {
        "tags": ["Sending"],
        "summary": "Browse the delivery log",
        "description": "Sent/failed/pending outbound messages, newest first. `total` is the count matching the filters.",
        "operationId": "listSent",
        "parameters": [
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Offset" },
          {
            "name": "status",
            "in": "query",
            "description": "Exact-match filter on delivery status.",
            "schema": { "$ref": "#/components/schemas/DeliveryStatus" }
          },
          {
            "name": "projectName",
            "in": "query",
            "description": "Exact-match filter on the sending project.",
            "schema": { "type": "string" }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Exact-match filter on the recipient number.",
            "schema": { "type": "string" }
          },
          { "$ref": "#/components/parameters/ModemFilter" }
        ],
        "responses": {
          "200": {
            "description": "A page of the delivery log.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "const": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "messages": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/SentMessage" }
                        },
                        "total": { "type": "integer", "description": "Count matching the filters." },
                        "limit": { "type": "integer" },
                        "offset": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "The gateway's `API_KEY`."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The gateway's `API_KEY` as a bearer token."
      }
    },
    "parameters": {
      "Limit": {
        "name": "limit",
        "in": "query",
        "description": "Page size (clamped to 1–500).",
        "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 500 }
      },
      "Offset": {
        "name": "offset",
        "in": "query",
        "description": "Rows to skip.",
        "schema": { "type": "integer", "default": 0, "minimum": 0 }
      },
      "ModemFilter": {
        "name": "modem",
        "in": "query",
        "description": "Restrict to one modem by id.",
        "schema": { "type": "string" }
      }
    },
    "schemas": {
      "DeliveryStatus": {
        "type": "string",
        "enum": ["pending", "retrying", "sent", "failed", "unconfirmed"],
        "description": "`pending` → queued; `retrying` → a send attempt failed, another is scheduled; `sent` → confirmed by the modem; `failed` → all attempts exhausted; `unconfirmed` → handed to the modem but the confirmation timed out (not retried — the SMS may already be delivered)."
      },
...
      "SentMessage": {
        "type": "object",
        "description": "A delivery-log entry.",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "timestamp": { "type": "string", "format": "date-time", "description": "When the send was queued." },
          "to": { "type": "string" },
          "message": { "type": "string" },
          "projectName": { "type": "string" },
          "ip": { "type": ["string", "null"], "description": "Client IP that queued the send." },
          "status": { "$ref": "#/components/schemas/DeliveryStatus" },
          "sentAt": { "type": ["string", "null"], "format": "date-time" },
          "reference": {
            "type": ["string", "null"],
            "description": "Modem's send reference (e.g. `+CMGS: 12` on serial)."
          },
          "modemId": {
            "type": ["string", "null"],
            "description": "The modem that carried (or last attempted) the message; `null` while the send is still pending on a round-robin job."
          },
          "error": { "type": ["string", "null"], "description": "Last modem error on `retrying`/`failed`." }
        },
        "example": {
          "id": "7edf5906-55c5-4b57-b437-1502167c3ed6",
          "timestamp": "2026-07-06T10:00:00.000Z",
          "to": "+99360123456",
          "message": "Hello",
          "projectName": "E-Center",
          "ip": "::1",
          "status": "sent",
          "sentAt": "2026-07-06T10:00:04.000Z",
          "reference": "+CMGS: 12",
          "modemId": "zte",
          "error": null
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "const": false },
          "message": { "type": "string" }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid API key.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "success": false, "message": "Unauthorized" }
          }
        }
      },
      "ServerError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  }
}